fix(plugin): resolve all compiler warnings with juce_recommended_warning_flags
Fix 190+ compiler warnings to achieve zero-warning builds with strict
JUCE warning flags enabled. Changes include:

- Add explicit static_cast<float>() for int-to-float conversions in
  graphics code (bounds, coordinates, progress bars)
- Add static_cast<size_t>() and static_cast<unsigned int>() for
  sign-conversion warnings in container operations
- Fix variable shadowing by renaming conflicting local variables
- Add missing enum cases to switch statements (FilterType::All)
- Rename shadowed lambda captures and constructor parameters

Warning categories fixed:
- -Wimplicit-int-float-conversion (100+ instances)
- -Wsign-conversion (60+ instances)
- -Wshadow-uncaptured-local (17 instances)
- -Wswitch-enum (5 instances)
- -Wshadow (5 instances)

+++ plugin/src/audio/BufferAudioPlayer.cpp
@@ -226,7 +226,8 @@ void BufferAudioPlayer::processBlock(juce::AudioBuffer<float> &buffer, int numSa
   }
 
   // Update position
-  juce::int64 newPos = static_cast<juce::int64>(currentPos + resamplingRatio * numSamples);
+  juce::int64 newPos =
+      static_cast<juce::int64>(static_cast<double>(currentPos) + resamplingRatio * static_cast<double>(numSamples));
   newPos = juce::jlimit<juce::int64>(0, totalSamples, newPos);
   currentSamplePosition = newPos;
